# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SADI API",
    "version": "1.0"
  },
  "paths": {
...
    "/unified/v1/wpa/powerbi/powerbi/dashboard/{dashboard_id}": {
      "get": {
        "tags": [
          "PowerBI-Style"
        ],
        "summary": "Get Dashboard",
        "description": "Returns full dashboard: layout, widgets, filters, metadata.",
        "operationId": "get_dashboard_unified_v1_wpa_powerbi_powerbi_dashboard__dashboard_id__get",
        "parameters": [
          {
            "name": "dashboard_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dashboard Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardConfig"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/unified/v1/wpa/powerbi/powerbi/data/query": {
      "post": {
        "tags": [
          "PowerBI-Style"
        ],
        "summary": "Query Data",
        "description": "Executes a dataset query (SQL-like, pandas-like, or connector).\nAutomatically handles:\n  - DBs\n  - CSV/Excel/Parquet\n  - Google Drive / OneDrive\n  - DataLake\n  - AutoML enriched data",
        "operationId": "query_data_unified_v1_wpa_powerbi_powerbi_data_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/unified/v1/wpa/powerbi/powerbi/data/widget/{dashboard_id}/{widget_id}": {
      "get": {
        "tags": [
          "PowerBI-Style"
        ],
        "summary": "Get Widget Data",
        "description": "Returns processed widget data (aggregations, group-bys, filters applied).",
        "operationId": "get_widget_data_unified_v1_wpa_powerbi_powerbi_data_widget__dashboard_id___widget_id__get",
        "parameters": [
          {
            "name": "dashboard_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dashboard Id"
            }
          },
          {
            "name": "widget_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Widget Id"
            }
          },
          {
            "name": "filter_values",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "JSON string of filters",
              "title": "Filter Values"
            },
            "description": "JSON string of filters"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/unified/v1/wpa/powerbi/powerbi/model/recommend": {
      "post": {
        "tags": [
          "PowerBI-Style"
        ],
        "summary": "Recommend Dashboard",
        "description": "AutoML suggestions:\n  - Best KPIs\n  - Best chart types\n  - Best variables to analyze\n  - Insights (correlations, trends)\n  - Feature importance\n  - Clusters / segments",
        "operationId": "recommend_dashboard_unified_v1_wpa_powerbi_powerbi_model_recommend_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DashboardConfig": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Dashboard ID"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Human title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category",
            "default": "general"
          },
          "global_filters": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Global Filters"
          },
          "layout": {
            "type": "object",
            "title": "Layout",
            "description": "Layout: {widgets: [...], grid: {...}}"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "layout"
        ],
        "title": "DashboardConfig"
      },
      "DataQueryRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "local"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "DataQueryRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RecommendationRequest": {
        "properties": {
          "dataset_name": {
            "type": "string",
            "title": "Dataset Name"
          },
          "sample_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Limit",
            "default": 5000
          }
        },
        "type": "object",
        "required": [
          "dataset_name"
        ],
        "title": "RecommendationRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/token"
          }
        }
      }
    }
  }
}